cross-validation split
The 20% of the train images of each class are used as validation. The
code makes a random selection of this 20% and copies the images to ta
validation folder. Also, it generates 4 .dat with the paths and labels
to the validation images and a new files to the new train images
removing the paths from the validation ones.